Restructure the key-related functions for usuability
